refactor(frontend): consolidate layout and card components
