+++ frontend/src/app/evolution/_components/index.ts
@@ -1,4 +1,2 @@
-import ChapterSummary from "./ChapterSummary";
-import TimelineItem from "./TimelineItem";
-
-export { ChapterSummary, TimelineItem };
+export { ChapterSummary } from "./ChapterSummary";
+export { TimelineItem } from "./TimelineItem";
