--- frontend/src/components/display/index.ts
@@ -1,3 +1,0 @@
-import BatchResultsDisplay from "./BatchResultsDisplay";
-
-export { BatchResultsDisplay };
