+++ frontend/src/components/form/FormField.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
-import FormField from "./FormField";
+import { FormField } from "./FormField";
 
 describe("FormField", () => {
 	it("renders label correctly", () => {
