+++ frontend/src/components/form/index.ts
@@ -1,17 +1,7 @@
-import BatchFormControls from "./BatchFormControls";
-import ColorContrastMessage from "./ColorContrastMessage";
-import ColorControls from "./ColorControls";
-import CoverForm from "./CoverForm";
-import CoverFormControls from "./CoverFormControls";
-import CoverPreviewDisplay from "./CoverPreviewDisplay";
-import FormField from "./FormField";
-
-export {
-	BatchFormControls,
-	ColorContrastMessage,
-	ColorControls,
-	CoverForm,
-	CoverFormControls,
-	CoverPreviewDisplay,
-	FormField,
-};
+export { BatchFormControls } from "./BatchFormControls";
+export { ColorContrastMessage } from "./ColorContrastMessage";
+export { ColorControls } from "./ColorControls";
+export { CoverForm } from "./CoverForm";
+export { CoverFormControls } from "./CoverFormControls";
+export { CoverPreviewDisplay } from "./CoverPreviewDisplay";
+export { FormField } from "./FormField";
