--- frontend/src/components/layout/Footer.test.tsx
@@ -1,25 +1,0 @@
-import { render, screen } from "@testing-library/react";
-import { describe, expect, it } from "vitest";
-import Footer from "./Footer";
-
-describe("Footer", () => {
-	it("renders copyright text", () => {
-		render(<Footer />);
-		const copyright = screen.getByText(/© \d{4} Victoria Cheng/i);
-		expect(copyright).toBeInTheDocument();
-	});
-
-	it("renders GitHub link with correct href and attributes", () => {
-		render(<Footer />);
-		const githubLink = screen.getByText(/GitHub/i).closest("a");
-		expect(githubLink).toBeInTheDocument();
-		expect(githubLink).toHaveAttribute("target", "_blank");
-		expect(githubLink).toHaveAttribute("rel", "noopener noreferrer");
-	});
-
-	it("applies correct classes for styling", () => {
-		const { container } = render(<Footer />);
-		const footer = container.querySelector("footer");
-		expect(footer).toHaveClass("bg-emerald-200", "text-gray-900", "py-4");
-	});
-});
