--- frontend/src/components/layout/MainLayout.tsx
@@ -1,22 +1,0 @@
-import type { ReactNode } from "react";
-import Footer from "./Footer";
-import Header from "./Header";
-
-interface MainLayoutProps {
-	children: ReactNode;
-}
-
-export default function MainLayout({ children }: MainLayoutProps) {
-	return (
-		<div className="flex flex-col min-h-screen bg-emerald-100 text-gray-900">
-			{/* Header */}
-			<Header />
-
-			{/* Main content */}
-			<main className="flex-1 w-[90%] max-w-7xl mx-auto py-8">{children}</main>
-
-			{/* Footer */}
-			<Footer />
-		</div>
-	);
-}
