+++ frontend/src/components/ui/Button.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render, screen } from "@testing-library/react";
 import { describe, expect, it, vi } from "vitest";
-import Button from "./Button";
+import { Button } from "./Button";
 
 describe("Button", () => {
 	it("renders children correctly", () => {
