+++ frontend/src/components/ui/Button.tsx
@@ -5,7 +5,7 @@ interface ButtonProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
 	isLoading?: boolean;
 }
 
-export default function Button({
+export function Button({
 	variant = "primary",
 	className,
 	disabled,
