+++ frontend/src/components/ui/Cards.tsx
@@ -1,3 +1,21 @@
+import { cn } from "@/lib";
+
+interface CardProps extends React.HTMLAttributes<HTMLDivElement> {}
+
+export function Card({ className, children, ...props }: CardProps) {
+	return (
+		<div
+			className={cn(
+				"rounded-xl border border-emerald-200 bg-gray-50 p-4 shadow-sm transition-all duration-200 hover:border-emerald-500 hover:shadow-md",
+				className,
+			)}
+			{...props}
+		>
+			{children}
+		</div>
+	);
+}
+
 interface KPICardProps {
 	title: string;
 	value: number | string;
@@ -25,7 +43,7 @@ const colorStyles: Record<string, { bg: string; text: string }> = {
 	white: { bg: "bg-white border border-gray-200", text: "text-gray-900" },
 };
 
-export default function KPICard({
+export function KPICard({
 	title,
 	value,
 	color,
