+++ frontend/src/components/ui/ColorPicker.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render, screen } from "@testing-library/react";
 import { describe, expect, it, vi } from "vitest";
-import ColorPicker from "./ColorPicker";
+import { ColorPicker } from "./ColorPicker";
 
 describe("ColorPicker", () => {
 	it("renders correctly", () => {
