+++ frontend/src/components/ui/SectionTitle.tsx
@@ -5,7 +5,7 @@ type SectionTitleProps = React.HTMLAttributes<HTMLHeadingElement> & {
 	as?: "h1" | "h2" | "h3" | "h4" | "h5" | "h6";
 };
 
-export default function SectionTitle({
+export function SectionTitle({
 	className,
 	size = "md",
 	as = "h2",
