+++ frontend/src/components/ui/Select.test.tsx
@@ -1,6 +1,6 @@
 import { fireEvent, render, screen } from "@testing-library/react";
 import { describe, expect, it, vi } from "vitest";
-import Select from "./Select";
+import { Select } from "./Select";
 
 describe("Select", () => {
 	it("renders correctly", () => {
