+++ frontend/src/components/ui/Skeleton.test.tsx
@@ -1,6 +1,6 @@
 import { render } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
-import Skeleton from "./Skeleton";
+import { Skeleton } from "./Skeleton";
 
 describe("Skeleton", () => {
 	it("renders correctly with default styles", () => {
