+++ frontend/src/components/ui/Skeleton.tsx
@@ -2,7 +2,7 @@ import { cn } from "@/lib";
 
 interface SkeletonProps extends React.HTMLAttributes<HTMLDivElement> {}
 
-export default function Skeleton({ className, ...props }: SkeletonProps) {
+export function Skeleton({ className, ...props }: SkeletonProps) {
 	return (
 		<div
 			className={cn("animate-pulse rounded-md bg-gray-200", className)}
